feat(fe): 블럭 크기조절, 이동 socket 버그수정
